Move scalePanel to shared folder.
