+++ assets/src/edit-story/elements/shared/scalePanel.js
